Create the custom Dropdown directive
Replace Bootstrap's default dropdown menu with a custom directive
implemented to handle the system menu at the top right corner of
the application.

This new directive makes use of HostListener to listen to mouse
events and also uses HostBinding to dynamically change the class
applied to the menu element, making it visible when the action
menu is clicked.