Create the RecipeService and ShoppingListService
Add two services to manage the data retrieval regarding recipes
and shopping list items. Such services provide methods that allow
the application to retrieve the desired information and can be
reused throughout different modules that need to perform them.

This is an example of dependency injection in Angular 2 and also
demonstrates the usage of services.