Create RecipeEditComponent, RecipeStartComponent and add routing
Implement two new components to allow the edition of an existing
recipe and to act as a landing page in the application (asking the
user to select a recipe from the list) and, after that, set up
routing to make the different application components to communicate
with each other and be visible in specific parts of the page.

This commit makes the app composed of multiple pages, allowing each
component to have its own page. Besides that, the app is now much
more functional and real than before.