Form submission and recipe ingredient management
Enable functionality for form submission so that changes in text
fields are supported, as long as the addition or removal of
ingredients from the list on each recipe.

The recipe component is now able to show, view and create new
recipes, and each recipe can now be edited and its ingredients
can be changed, added or removed, as expected.