Implement Http requests to handle app information
HTTP requests are now used to store and retrieve the recipes from
a database, allowing the application details to be persisted so
recipes can be kept throughout different sessions and not get
erased when a page reload happens, for instance.

For this simple example, Firebase is used as the database for
persisting the recipes JSON objects.