Auto-load recipes on start up
A small hook was implemented to allow the persisted recipes to get
loaded from Firebase at the application start up. When the recipe
service's getRecipes() is called for the first time, it will load
the entries from database so the array is correctly populated.

As this method now works totally integrated with Firebase, the fake
(mocked) recipes were removed from the app. Now it works 100% with
real information present in the database.