Use feature modules and lazy loading of routes
Introduce two feature modules to aggregate the content of Recipes
and Shopping List modules, respectively, so the main AppModule can
be decluttered and code is less coupled (and messy).

Also introduce lazy loading of routes to allow the modules to be
more independent. This also allows performance to have be slightly
improved by only loading what is necessary for the current app
usage.